Version 1.0 functionality:  Fixing issues with displaying quiz multi-choice images in quiz and review modes.

+++ qpalx-elearning-web-zone-content-admin/src/main/java/com/quaza/solutions/qpalx/elearning/web/zone/content/admin/curriculum/QPalXMicroLessonAdminController.java
@@ -225,7 +225,9 @@ public class QPalXMicroLessonAdminController {
             iRedirectStrategyExecutor.sendRedirectWithError(targetURL, error, WebOperationErrorAttributesE.Invalid_Delete_Operation, request, response);
         } else {
             iqPalXEMicroLessonService.delete(qPalXEMicroLesson);
-            ieLearningStaticContentService.deleteELearningMediaContent(qPalXEMicroLesson.geteLearningMediaContent());
+            if (qPalXEMicroLesson.geteLearningMediaContent() != null) {
+                ieLearningStaticContentService.deleteELearningMediaContent(qPalXEMicroLesson.geteLearningMediaContent());
+            }
             String targetURL = "/view-admin-qpalx-micro-elessons?qpalxELessonID=" + qPalXELesson.getId();
             iRedirectStrategyExecutor.sendRedirect(request, response, targetURL);
         }
